requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
attrs=21.4.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
bleach=5.0.0=pypi_0
ca-certificates=2021.10.26=h06a4308_2
cached-property=1.5.2=pypi_0
cachetools=5.0.0=pypi_0
certifi=2021.10.8=py37h06a4308_2
cffi=1.15.0=pypi_0
charset-normalizer=2.0.12=pypi_0
communities=3.0.0=pypi_0
cppimport=21.3.7=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
debugpy=1.6.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dgl-cu113=0.7.2=pypi_0
easydict=1.9=pypi_0
entrypoints=0.4=pypi_0
faiss-gpu=1.7.2=py3.7_h28a55e0_0_cuda11.3
fastjsonschema=2.15.3=pypi_0
fonttools=4.32.0=pypi_0
google-auth=2.6.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.43.0=pypi_0
h5py=3.6.0=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.11.1=pypi_0
importlib-resources=5.7.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.13.0=pypi_0
ipython=7.32.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.7.0=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.1=pypi_0
joblib=1.1.0=pypi_0
jsonschema=4.4.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.3.0=pypi_0
jupyter-console=6.4.3=pypi_0
jupyter-core=4.10.0=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=1.1.0=pypi_0
kiwisolver=1.4.2=pypi_0
ld_impl_linux-64=2.35.1=h7274673_9
libfaiss=1.7.2=hfc2d529_0_cuda11.3
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libstdcxx-ng=9.3.0=hd4cf53a_17
llvmlite=0.38.0=pypi_0
mako=1.2.0=pypi_0
markdown=3.3.6=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
nbclient=0.6.0=pypi_0
nbconvert=6.5.0=pypi_0
nbformat=5.3.0=pypi_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.5=pypi_0
networkx=2.6.3=pypi_0
notebook=6.4.11=pypi_0
numba=0.55.1=pypi_0
numpy=1.21.5=pypi_0
numpy-base=1.21.2=py37h79a1101_0
oauthlib=3.2.0=pypi_0
openssl=1.1.1m=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.3.5=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.1=pypi_0
pip=21.2.2=py37h06a4308_0
prometheus-client=0.14.1=pypi_0
prompt-toolkit=3.0.29=pypi_0
protobuf=3.19.4=pypi_0
psutil=5.9.0=pypi_0
ptyprocess=0.7.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pybind11=2.9.1=pypi_0
pycparser=2.21=pypi_0
pygments=2.12.0=pypi_0
pyparsing=3.0.8=pypi_0
pyrsistent=0.18.1=pypi_0
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pypi_0
pytz=2021.3=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=22.3.0=pypi_0
qtconsole=5.3.0=pypi_0
qtpy=2.0.1=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.11.2=pypi_0
send2trash=1.8.0=pypi_0
setproctitle=1.2.3=pypi_0
setuptools=58.0.4=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
soupsieve=2.3.2.post1=pypi_0
sqlite=3.37.2=hc218d9a_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=2.5=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.13.3=pypi_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.1.1=pypi_0
tk=8.6.11=h1ccaba5_0
torch=1.10.2+cu102=pypi_0
torch-tb-profiler=0.3.1=pypi_0
torchaudio=0.10.2+cu102=pypi_0
torchvision=0.11.3+cu102=pypi_0
tornado=6.1=pypi_0
tqdm=4.63.0=pypi_0
traitlets=5.1.1=pypi_0
typing-extensions=4.1.1=pypi_0
urllib3=1.26.8=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.0.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.6.0=pypi_0
xz=5.2.5=h7b6447c_0
zipp=3.7.0=pypi_0
zlib=1.2.11=h7f8727e_4
